refactor(gh-pr-release-tagger): extract extractOriginalPRNumber to dep-free module
The spec file pulls this helper in via the task index, which transitively
loads @/src/db and spins up a real Mongo client at import time. Moving the
parser to a standalone module lets unit tests import it directly without
touching the DB layer or the GitHub client wrapper.

Addresses Copilot comments 3064000747 and 3065978636.
